refactor: split job domain from main.rs
